Add scripts module
This module enhances the use of python scripts by allowing the use of
pipenv environments. It interfaces using a cli call, and provides an
Argparse parser to receive the same standard inputs as typically
available in a native Snakemake `script`. Also provides a `ScriptDict`
class to more easily include elements in your CLI call (e.g. `input`,
`output` etc). Without this, when including dicts, the names of the
dict keys must be provided as arguments to `pyscript`